GH-131 Move helper functions to separate files & namespace them
